+++ jitsi-meet/jitsi-meet.d/local/share/cook/bin/configure-prosody.sh
@@ -35,6 +35,7 @@ sep=$'\001'
 < "$TEMPLATEPATH/prosody.cfg.lua.in" \
   sed "s${sep}%%domain%%${sep}$DOMAIN${sep}g" | \
   sed "s${sep}%%email%%${sep}$EMAIL${sep}g" | \
+  sed "s${sep}%%turnpassword%%${sep}$HASHTURNPASSWORD${sep}g" | \
   sed "s${sep}%%keypassword%%${sep}$KEYPASSWORD${sep}g" \
   > /usr/local/etc/prosody/prosody.cfg.lua
 
@@ -43,6 +44,8 @@ sep=$'\001'
 echo -ne '\n\n\n\n\n\n\n\n\n\n\n' | prosodyctl cert generate "$DOMAIN" || true
 # shellcheck disable=SC2039
 echo -ne '\n\n\n\n\n\n\n\n\n\n\n' | prosodyctl cert generate auth."$DOMAIN" || true
+# shellcheck disable=SC2039
+echo -ne '\n\n\n\n\n\n\n\n\n\n\n' | prosodyctl cert generate localhost || true
 
 # Set up truststore
 keytool \