+++ jitsi-meet/jitsi-meet.d/local/share/cook/bin/configure-turnserver.sh
@@ -18,8 +18,9 @@ TEMPLATEPATH=$(dirname "$SCRIPT")/../templates
 # safe(r) separator for sed
 sep=$'\001'
 
-TURNPASSWORD=$(date | md5sum)
-HASHTURNPASSWORD=$(/usr/local/bin/turnadmin -P -p "$TURNPASSWORD")
+# moved to cook script so HASHTURNPASSWORD variable is available to prosody
+#TURNPASSWORD=$(date | md5sum)
+#HASHTURNPASSWORD=$(/usr/local/bin/turnadmin -P -p "$TURNPASSWORD")
 
 # copy over turnserver.conf
 < "$TEMPLATEPATH/turnserver.conf.in" \