+++ jitsi-meet/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Version bump for new base image
 * Implement websockets and turnserver
+* Script error fixes with introduction turnserver
 
 ---